Player Spawn Location
The spawn lacation of players is now fixed. It is random, but player will never even with the edge of the sprite appear out of the game area and it will also never touch another player. Also players are somewhat equally spaced out, so they don't all spawn in one fifteenth of the game area.